requirements.txt
# Core ML
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Hugging Face ecosystem
transformers>=4.36.0
datasets>=2.16.0
peft>=0.7.0
accelerate>=0.25.0
timm>=0.9.12

# Data handling
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
scikit-learn>=1.3.0
roboflow>=1.1.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Grad-CAM
pytorch-grad-cam>=1.5.0

# Quantization
bitsandbytes>=0.41.0; sys_platform != 'win32'
optimum>=1.14.0

# Environment
python-dotenv>=1.0.0

# Deployment
gradio>=4.10.0

# Metrics
torchmetrics>=1.2.0

# Notebook support
ipykernel>=6.25.0
ipywidgets>=8.1.0
tqdm>=4.66.0

# Optional: experiment tracking
# wandb>=0.16.0
# tensorboard>=2.15.0
